ADA_Project/backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
openpyxl==3.1.5
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.53.0
pydantic==2.9.2
werkzeug==3.0.4